return categories as list column
Return all categories as one list column instead of up to five categories in five columns named category.1 to 5. This not only preserves all information (category 6 is left out by now), but also makes it easier to work with the data. Unnesting the column is as simply as `tidyr::unnest(mydataframe)`. One might of course debate if this structures is still tidy...